Create README.md
Over the past month, I had the opportunity to work on a variety of interesting tasks that not only challenged me but also helped me grow as a developer. Here are some of the projects I worked on:

#Number Guessing 🎲

Developed a Java program that allows users to play a number guessing game. Implemented logic to generate a random number, receive user inputs, and provide feedback based on the guesses. This task improved my understanding of conditional statements and user input handling.

#Word Counter 📝

Created a Java program to count the number of words in a given text file. Utilized file handling techniques to read the input file and implemented an algorithm to count the words. This project enhanced my file handling skills and introduced me to string manipulation in Java.

#ATM Interface 💳

Designed an interactive ATM interface using Java Swing. Implemented features like balance inquiry, cash withdrawal, and deposit functions. This project provided hands-on experience with graphical user interface (GUI) development and event handling.